+++ vue/components/ui/organisms/index.js
@@ -1,4 +1,5 @@
 import { Chat } from "./chat/chat.vue";
+import { ContainerForm } from "./container-form/container-form.vue";
 import { Details } from "./details/details.vue";
 import { EntityList } from "./entity-list/entity-list.vue";
 import { EntityShow } from "./entity-show/entity-show.vue";
@@ -10,6 +11,7 @@ import { TableMenu } from "./table-menu/table-menu.vue";
 
 const install = Vue => {
     Vue.component("chat", Chat);
+    Vue.component("container-form", ContainerForm);
     Vue.component("details-ripe", Details);
     Vue.component("entity-list", EntityList);
     Vue.component("entity-show", EntityShow);
@@ -20,6 +22,6 @@ const install = Vue => {
     Vue.component("table-menu", TableMenu);
 };
 
-export { Chat, Details, EntityList, EntityShow, Form, Header, Listing, TableMenu };
+export { Chat, ContainerForm, Details, EntityList, EntityShow, Form, Header, Listing, TableMenu };
 
 export default install;
